# najah-delivery-os/services/express-api/openapi.yaml
openapi: 3.0.3
info:
  title: Najah Delivery API
  description: Express API service for Najah Delivery OS - providing comprehensive delivery management endpoints
  version: 1.0.0
  contact:
    name: Najah Delivery Support
    email: support@najahdelivery.com
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.najahdelivery.com/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /tracking/update:
    post:
      tags:
        - Tracking
      summary: Update order tracking
      description: Update order and courier location/status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - orderId
                - courierId
                - location
              properties:
                orderId:
                  type: string
                courierId:
                  type: string
                location:
                  type: object
                  properties:
                    coordinates:
                      type: array
                      items:
                        type: number
                status:
                  type: string
      responses:
        '200':
          description: Tracking updated
  /tracking/{orderId}:
    get:
      tags:
        - Tracking
      summary: Get order tracking info
      description: Public endpoint to track order status
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tracking information
...